scripts/EMAGE_2024/requirements.txt
ConfigArgParse==1.7
fasttext==0.9.2
h5py==3.10.0
imageio==2.31.4
ipython==8.12.3
joblib==1.3.2
librosa==0.10.1
lmdb==1.4.1
loguru==0.7.2
matplotlib==3.7.3
moviepy==1.0.3
numpy
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
pandas==1.5.3
peakutils==1.3.4
ptflops==0.7.1.2
pyarrow==0.16.0
python_igraph==0.11.3
pytorch3d
pyvirtualdisplay==3.0
PyYAML==6.0.1
replicate==0.15.4
scikit_learn==1.3.2
scipy
soundfile==0.12.1
termcolor==2.4.0
textgrid==1.5
torch==2.1.0
torchvision
tqdm==4.66.1
transformers==4.35.2
trimesh==3.23.5
wandb==0.16.0
smplx
tensorboard